ckssh-add: Use config file keys if not specified on command line
We interpolate `~` in the config file as `$HOME`, but this will
break if `$HOME` has a colon in it. (This is considered unlikely
enough that we'll live with the problem for now.)

This code is probably not entirely optimal, but is stretching
my brain after many hours of work today, so let's optimize
later.