Make ssh command work
The new `print0_ssh_args` function is the core of this, which builds a
new command line from the supplied command line args, the ckssh config
file args for that host and compartment, and then the host and
remaining args. We build a command line of NULL-separated arguments
that we then pass to `xargs -0` as a clever way of avoiding another
evaluation by the shell of the already-evaluated args.

`scp` is still broken due to differences in the command line args
(hostname:filename); we add the symlink and test for usage message on
no args, but otherwise it's unimplemented.

This brought in a fair amount of semi-duplicate code from other
functions that needs to be refactored.