bin/ckssh: Fix what we thought was weird need for a colon
In the first use the `:` somehow fixed that we did not quote the
`$(...)` with a pipe and while loop in it.

I dunno, it's all mysterious shell quoting shit. Best not to think
about it too hard lest your brain explode, leaving you unable to
program ever again except perhaps in languages like COBOL.