ckssh: We need to quote all uses of $(...) that could have spaces
This is mainly related to pathnames, where we might have a space in
the path or filenames.

However, there was also a mysterious issue where we had to put
a `:` in a while loop that turned out to be due to the lack of
quoting as well, as the while loop was in a $(...) construct.