Implement load_compartment_config for future use
This loads the compartment configuration into an array.

It's somewhat inefficient as we read the config twice, once to generate
the return value if there's a failure, and again when we've proven it
can't fail. This is because we use a subshell on the call that actually
reads the data and inserts it into the array.

We no longer need the print_config array assignment test as
load_compartment_config now serves as the example of how to do this.