Implement load_host_config for future use
This appends the host config, minus any `CK_CompartmentName` directives,
to the given array, and then calls `load_compartment_config` to append
the compartment configuration to that array.

This includes twice any global configuration commands (once from the
`CK_Host` scan and once again from the `CK_Compartment` scan, but as
this is harmless we don't try to fix it.