ckssh-add: Take only compartment names now
We cannot take a host name when hosts are scanned before compartments
because then a host or host wildcard might match a compartment name,
rendering us unable to ckssh-add the compartment itself.

We could scan compartments before hosts just for this command, but
that's a bit odd since it's the reverse of how we do it everywhere
else, and it just doesn't seem a very useful feature anyway.