Proof of concept that we can have separate output stream for eval
We want `ckssh.py` to be able to print to stdout, print to stderr, and
entirely independently of that return shell commands to the caller for
it to execute. (This is so that it can, e.g., set environment
variables in the calling shell's environment.)

To do this we offer an `--eval-file` option to which ckssh.py can
write code it would like the caller to execute. (If not provided, the
code is written to stdout.)

The `--proof-of-concept` opton shows that this works by printing text
to both stdout and stderr and printing an `echo` command and
environment variable setting to the eval-file which is then executed
by the `ckset()` shell function.