Use subcommand arguments for bash-init and proof-of-concept
We now define a generic shell function `ckcommand` to call `ckssh.py`,
set up the temp file for eval, execute output, etc., and `ckset`
becomes a simple wrapper that calls `ckcommand ckset` (which does not
work yet due to lack of a `ckset` subcommand).

Change `--bash-init` and `--proof-of-concept` to be subcommands, which
gets rid of the annoying "you could specify both or neither" problem.
We use a quick hack for the moment to figure out what function to call
for each subcommand, but note that later we should probably switch to
proper subcommand parsers.