Differentiate between unknown and nonexistent compartment
`CK.compartment_name()` now `None` if $SSH_AUTH_SOCK is not set (or
empty); `CK.UnknownCompartment` if it's set but not to the path to the
socket of a compartment we know; otherwise the name.

The `ckset` print routine now uses this to print 'No compartment.' or
'Unknown compartment.' to stderr, or the compartment name to stdout.
(The README has been updated to document this API.) We still do not
exit with a code based on whether or not an agent is available,
however. (And one not being available may not be an actual error if
it's a known compartment and we could start one.)

Later we can (given `-v`) print out the list of keys in an unknown
compartment, perhaps including information about what known
compartments use those keys, or even determining that the unknown
compartment is configured the same as a known one.

It's an open question whether one that matches a known compartment
should be treated as that known compartment. That would be convenient
when SSH'ing into a remote host, but if it's not really the known
compartment at the source, extra keys might be added.