CK: Replace compartment_name() with compartment_from_sock()
Moving forward we need a way to get further information about the
compartment, so we want to return the CompartmentConfig object
instead of just the name, and let the clients query that object
for whatever info they need.

We also move CompartmentConfig into the CK class since it's tightly
bound to it.