Add some better exit codes instead of always exiting with 0
The functional test now picks an SSH_AUTH_SOCK path more certain not
to exist and checks that ckset exits 1 to indicate no agent is running
there. (But actually it still exits 1 even if an agent is running.)

We now print an error and exit 2 if additional ckset args are given,
since we can't handle them yet.

Also we make pytest a little bit quieter because the output was making
it too easy to miss 'FAILED!' at the end (generated by the trap) when
the Test script didn't work properly.