Check to see if compartment (agent) is running
We now have the correct exit code for when a compartment is or isn't
running (i.e., there is an SSH agent server running on the socket).

It's perhaps a bit hokey to call out to `ssh-add -l` and use its
return code, but it's not clear yet if we want to introduce a
dependency on a library to talk to the agent directly, or perhaps just
write our own (hopefully almost trivial) code to message the agent.