Fix parsing bug that was dropping compartments
Immediately on encountering a compartment delcaration we want to add
it to the compartments list; we can still mutate it after it's added
to add keyfiles, etc.

This avoids the problem where we wouldn't add a compartment to the
list if it wasn't followed by another compartment or `CK_Host`
declaration.