Add CK_confirm (yes/no) to pass `-c` option to ssh-add
This defaults to yes, making the user confirm each use of the key (so
long as the agent supports it). We prefer to default to the more secure
option when not told explicitly what to use.

When confirmation is enabled, the key add operation should probably
check for the presence of `ssh-askpass` and warn if it's not present,
since otherwise the OpenSSH agent will always refuse to sign.

The parser for this is starting to get a bit messy, with repetitive
code. We need to figure a better way to sort this out.