Extract environment as parameter to subcommands
The subcommands now all take (args, env) as parameters which makes
functional testing easier and lets us fix the environment-dependent
failure of test_ckset_show().

This involved reworking environment access in some other functions:
* runtimedir() no longer has a default of os.environ for the
  environment parameter
* CK() now requires an environment and no longer uses os.environ

We also found that `from sys import stderr` is not compatible with
pytest's capsys fixture because that takes a reference to the original
stderr IO object before it's replaced by capsys, and so capsys doesn't
capture the output. We changed the one spot where this caused trouble
to use sys.stderr instead of the local stderr, but this requires
further cleanup.