t/functional: Test ckset show with unknown and known compartments
The little hack to make expanduser() use $HOME from the passed-in
environment is not pleasant, but since expanduser() relies on a global
(os.environ), we're kinda stuck with it. Hopefully this won't have
unfortunate knock-on effects on other parts of the program or tests.

Maybe we should set os.environ directly in the functional tests,
instead?