WIP tox.ini: Try using tox to run our tests in multiple virtualenvs
This is a rough sketch, and happily exposed a bug (when $HOME is not
set in os.environ nor passed into to CK()'s env).

skipsdist=true avoids trying to run a setup.py file, since we neither
have nor need one.

This doesn't run all the tests, just the pytest ones. If we were going
to use this for real we probably need to have the top-level Test
script handle setup etc. of tox and move the running of pytest stuff
and Bash functional tests into their own subscript under t/ for tox to
run.

It looks like the "normal" way to run tox is to ask the developer to
install the package and run `tox` at the command line. If we don't
want to do that we may want to keep our own virtualenv in which to
install it, or perhaps we can do a bootstrap install the way we do
with virtualenv itself.