t/functional: Rework how we call the functions under test
Instead of hacking our own parsed args object, we use ckssh's argument
parser. We duplicate the subcommand lookup rather than using it
directly because ckssh's main() does a sys.ext(); we should look for a
way to reduce this duplication.

This new method also consolidates most of the (fake) environment
settings in one place. This means we're now running with a test
environment with $HOME and $XDG_RUNTIME_DIR settings when we weren't
before, but this should be ok, at least for the moment.