t/unit: Reduce testing done by test_conf_with_config; add docstrings
test_parseconfig() tests that a full config file is parsed into a list
of compartments with all proper attributes, so a CK object really
just handles lookups into that list. Thus, we can just check the name
of the compartment returned and assume that any other problems will
be caught by test_parseconfig().

We also add docstrings to parseconfig() and CK to document the
responsibility split described above.