Change `-a` option to `-n` option
This is the simplest possible change we can do to implement this:
invert the sense of the tests that were checking `-a` and add `-n`
where we now need it.

Some rather more extensive addition of tests and refactoring is
required to bring us closer to the new spec, but that will be another
commit.