Add a `--version` option
This is mainly for the use of developers to make sure that whatever
bash function they're running is in turn running the development
version of ckssh rather than their standard one. (It shortcuts any
other actions.)

And yes, this program is starting to become a nasty, non-orthogonal
mess of special cases. Hopefully we can work our way out of this.