addkeys(): Return 2 if loaded_keynames is None
loaded_keynames is the result of fetch_keynames() and so instead of a
list we get None if the agent isn't running. Return 2 in this case to
match what ssh-add does, rather than throwing an exception.

This has no automated tests at the moment, either. We may need to do
some serious rework to get this into a more easily testable state.