Test: Switch to Python 2
Since we're aiming solely at Unix for the moment and aren't
anticipating the need for anything special from Python 3, let's
see if we can keep this working with Python 2 as well.

We do need to extent our test framework a bit further at some
point so that we test under 2 and 3, though.