ckset.py: Hack to handle compartment arg and starting an agent
The ability to say `ckssh foo` to start/switch to another compartment
hadn't yet been added to this rewrite branch. But I need this now, and
so it's hacked in as best I can with a lot of manual testing and not
really any unit testing.

Before continuing further with this, we should compare this with
the `rewrite2` branch and see which one we want to move forward with.

There are some later Python 3 idioms appearing here, such as f-strings;
this is fine as ckssh needs to run only on the machine running the agents,
presumably a modern system, and so doesn't need compatibility with the
old Pythons that may be on the hosts you're logging into.

BUG (previously existing, but annoying): ckset -v displays only known
keys for that configuration that are loaded in the agent, leaving out
any keys in the agent that are not known. This is at least confusing
and at worst a security issue. (E.g., someone ran `ssh-add` in the
wrong context.)