ckssh.py: ckcommand() now uses `source` instead of `eval`
`eval $(cat file)` doesn't work on multi-line output since all the
lines are concatenated to one; simply doing a `source file` fixes
this. (It could also be fixed by adding semicolons to the end of each
line written to eval-file, but that's more complex and doesn't provide
any gain.)

`eval $(ckssh.py bash-init)` does work because, though it's multi-line
output, it does use semicolon separation. This can't so easily be
changed however because `source` works only on files and we don't want
to have to make the user create a temp file to do the bash-init.