requirements.txt,pactivate: Update virtualenv build script and dependencies
The old `activate` script was an ad-hoc version of `pactivate`,
and so is replaced by that.

All the old dependencies, including on an ancient version of pytest,
are replaced by just pytest. Other things were either locked pytest
dependencies that shouldn't have been locked or were used back when
we still supported Python 2.x, which we no longer do.

Two tests are failing, but these were also failing before this update.
- t/functional.py::test_ckset_show_compartment_name
- t/functional.py::test_ckset_set_compartment_name_badargs