Test: Improve Bash error checks and failures
(Tests were broken before this change; this is the start of fixing them.)

Set -u and -E. Change trap from anything to just ERR, so we need not
unset the trap when done, and make it catch internal errors only.
Failures are now signaled with the fail() function.

We can clean with -c as well as -C now.