*: Fix failing pytest functional test with new XDG_RUNTIME_DIR
XDG_RUNTIME_DIR is a user-owned directory for sockets and the like.
Various tests seemed to be failing because it was set to
/test-xdg-rtdir-nonexistent, which of course doesn't exist, but it was
trying to write the sockets under that anyway. I'm not sure what was
previously going on with that or how I was testing with that before,
but it's simple enough to set it to $PROJDIR/.build/run/ for for the
entire test run, which will help make sure we never stomp on the
user's running configuration.

With that and fixing a few output string assertions that were
presumably from an earlier age (or with verbosity turned off) or
something like that, All tests run by pytest now pass.

So we're now failing at `ckcommand shell-interface-test`, which will
be the next thing to work on.